book circularity
