+++ src/format/epub/format-epub.ts
@@ -13,7 +13,7 @@ import {
   bookConfig,
   kBookCoverImage,
 } from "../../project/types/book/book-config.ts";
-import { BookExtension } from "../../project/types/book/book-extension.ts";
+import { BookExtension } from "../../project/types/book/book-shared.ts";
 import { createEbookFormat } from "../formats.ts";
 
 export function epubFormat(): Format {
