+++ src/project/types/book/book-crossrefs.ts
@@ -26,7 +26,7 @@ import { WebsiteProjectOutputFile } from "../website/website.ts";
 
 import { inputTargetIndex } from "../../project-index.ts";
 import { bookConfigRenderItems } from "./book-config.ts";
-import { isMultiFileBookFormat } from "./book-extension.ts";
+import { isMultiFileBookFormat } from "./book-shared.ts";
 
 export async function bookCrossrefsPostRender(
   context: ProjectContext,
