+++ src/project/types/book/book-extension.ts
@@ -24,29 +24,7 @@ import { inputTargetIndex } from "../../project-index.ts";
 import { bookConfigRenderItems } from "./book-config.ts";
 import { BookRenderItem } from "./book-config.ts";
 import { isHtmlOutput } from "../../../config/format.ts";
-
-export interface BookExtension {
-  // bool extensions are single file by default but can elect to be multi file
-  multiFile?: boolean;
-
-  // book extensions can modify the format before render
-  onSingleFilePreRender?: (format: Format, config?: ProjectConfig) => Format;
-
-  // book extensions can post-process the final rendered file
-  onSingleFilePostRender?: (
-    project: ProjectContext,
-    file: RenderedFile,
-  ) => void;
-}
-
-export function isMultiFileBookFormat(format: Format) {
-  const extension = format.extensions?.book as BookExtension;
-  if (extension) {
-    return extension.multiFile;
-  } else {
-    return false;
-  }
-}
+import { BookExtension, isMultiFileBookFormat } from "./book-shared.ts";
 
 export function onSingleFileBookPreRender(
   format: Format,
